Implemented a mechanism to generate Jax variables without having variables.
This is basically a simple substitute class that can be used instead of a full jax variable.
It is mostly usefull for creating arrays during testing.
However, it should also be used to cretae variables for which we do not have anything.

This essentially replaces the flags that allowed to explicitly specify shape and dtype in the `add_array()` function.