First round: Updated the naming of the subtranslator interface to the one Enrique suggested.
I have to say that I like it more than the old one, but I have not changed the terminology yet.
Thus I still use the world subtranslator and I think we should keep it because I still think that it is the best name.

However, I do not like the name because it suggests that it is able to already translate everything.