Updated the handling of JaCe variables.
Updated how the `__hash__` and `__eq__` works.
Before they only considered the name, which was not good.
There was also a short period, where all members were considered.
However, this commit changes them such that they only consider the `id()` value.
This is more in line with what Jax is doing.

Furthrmore if the name of a `JaCeVar` is empyt then the toolchain will consider them more like Jax variables.