Reworked how compilation works.
This commit removes the `csdfg` member of the `TranslatedJaxprSDFG` object, thus it is not cached anymore.
More importantly however, is the changes made to the compilation itself.
Before the SDFG was modified (global variables, arg_names) and then compiled, and then restored.
However, since the SDFG is part of teh CompiledSDFG this is inconsistent, for that reason the SDFG is deepcopied before processing.
This is surely not the best solution, but should prevent some nasty bugs.