This commit is the second part of the effort to make the list of currently active translators global.
The previous commit changed the driver to use them and adds the necessary infrastructure of managing them.
As before we want that in code such as:
```python
       @jit
       def foo(...):
              ...
       foo1 = foo.lower(...args1)
       # Modify the list of internal translators, i.e. call `add_subtranslator()`
       foo2 = foo.lower(...args2)
```
the second call to `lower()` uses the same translators as the first one.
What sounds simple is in fact relatively hard.
And we want that, in codes as
```
```python
       def foo(...):
              ...
       foo1 = jace.jit(foo)
       # Modify the list of internal translators, i.e. call `add_subtranslator()`
       foo2 = jace.jit(foo)
```
`foo1` and `foo2` use different translators as they where constructed differently.
As it is outlined inside in `JaceWrapped.__init__()`  the feature of "being able to manipulate the set of active translators from the outside of a `JaceWrapped` object" is not really useful and we thus forbid it.
As it is further outlined we _must_ make a copy (shallow copy is enough because the translators themselves are immutable) of the list of primitive translators that should be used, somewhere during the construction of a `JaceWrapped` object.
Which we do, for many reasons inside `JaceWrapped.__init__()`.

To include the set of used translators in the caching we use the address of the internal set of translators of the `JaceWrapped` object.
We can do this because we have by construction an immutable set, since we copied it at construction.
As an optimization, before we copy the set in `JaceWrapped.__init__()` we check if the set is the same as the global one (that is immutable).
This means that in situations such as:
```python
       def foo(A):
              ...
       foo1 = jace.jit(foo)(A)
       A[1, 2] = 0
       foo2 = jace.jit(foo)(A)
```
the code is only lowered once (this was not an intended design but a by product).
This captures the main use case in which the translators are set at the beginning and then never again.

We also included the Jax options to `jit`.
Currently we only really handle the case of no options, which is probably okay.