This commit partially add support for a global list of primitive translators.
This was Enrique's idea and it is really good, but it adds a lot of complexity.
For several reasons (mainly open PR) the commit is split into two parts one that can is merged into the PR (this one) and once that contains the parts that are development only.
Essentially we want allow code such as:
```python
       @jit
       def foo(...):
              ...
       foo1 = foo.lower(...args1)
       # Modify the list of internal translators, i.e. call `add_subtranslator()`
       foo2 = foo.lower(...args2)
```
because the list of translators is an implicit argument to the `jit` decorator, we expect that `foo2` is generated with the same translators as was `foo1`, see the next commit for a full description.
This commit essentially adds two things:
- the global list that stores the translator instances
- makes all necessary changes such that stuff works.

The managing of the global list is implemented such that it results in an immutable object.
Thus every time it is mutated a new list is created.
It might be a bit strange to do that, but the list is not changed that frequently (except upon loading) and it is only a shallow copy, since the translators are immutable themselves.
This approach allows some nice optimization further down.

For more information on the code see `jace.translator.managing.add_subtranslators`.