Started to remove some sources of cyclic imports.
According to [SE](https://stackoverflow.com/a/746067) the bst way to avoid that is using the `import ...` instaxt instead of the `from ... import` syntax.
Since the util module is accessed by almost everyone, it makes sense to use that stile inside it.
However, the translator package would also make much sense.

I am aware that this violates the codeing guide but in this case i do not care.