Updated the `JaxprTranslatorDriver::add_jax_name_mapping()` function.
Before, the behaviour was that if the mapping was already known but formally did not change no error was generated.
However, this has some issues when looking at nested stuff, so I promoted it to an error.

It would be possible to generate a more specific error but currently not.