Dynamic slicing can not run with disabled optimization.
The reason for this is quite interesting actually and there are several things that interact together to create the bug.
One part of the problem is, that we do not use scalars but only arrays and one part is probably a bug inside the code generator.
With enabled optimizations, start indexes of the dynamic slices are turned into symbols, which are implemented as (C) scalars.
However, if optimizations are disabled the start index is stored inside an array with length one.
In the Memlet, that we use to extract we essentially write something like `iteration_variable + start_index` which implies that `start_index` is a scalar, and if such an SDFG is compiled, then the code generator will generate `a[iteration_variable + start_index]`, but this is a compilation error, since `start_index` is a pointer and not a scalar.
The simple fix just to write `iteration_variable + start_index[0]` also does not work, as this seems to violate SDFG semantic because the code generator generates something like `a[iteration_variable + start_index(0)]` which is an error because `start_index` is obvious not a function.

Because there is no simple solution, this commit disables the tests for the slicing as a temporary solution.
The proper solution, which I will now implement, is to support scalars inside the translator.
In fact this will solve a lot of other problems.