Added a layer to handle all outputs and inputs.
This should mostly finalize the work needed to enable the distinction between arrays and scalars inside the builder. While there is now this distinction inside the builder and also in the input layer, i.e. there is no silent
convertion from scalar to array, at least on CPU, on GPU we will have to do that, but we are not there yet, we still return arrays in any case. This is because
We do this, because it is a limitation in DaCe itself, and Jax itself only returns arrays.