Integrated part of teh development branch into the new PR branch.
Essentially it is a partially copy of `src/jace` from the development branch (027ae355f38ed) to this branch.
However, the translators were not copied, thus they are still WIP mode.
Furthermore, all changes to the test were made such that they pass, i.e. they are in WIP mode.