Specified that a transient is exactly written once.
This si not fully true, actually there is onyl one access node that writes to the transient.
This is needed to implement primitives such as `concatenation`.
However, by a special prefix a variable can be written to multiple time, this is needed to implement the scans.
However, it might be possiblt to avoid that.