feat: Finalizing the initial infrastructure of JaCe (#18)
This PR introduces some missing pieces into mainline JaCe.

While the changes are very related to each other, the features this PR
introduces were not split.
A short summary of what it introduces:
- The stages now support pytrees as input and output.
- Inside the translators scalars and arrays are now distinguished.
- Support for arrays that are not in continuous C order.
- Type annotation in the wrapped objects (there are technical
limitations, see note in `src/jace/stages.py` for more).
- General changes in the organization of the code.
- Possibility to globally control the optimization levels (currently not
that useful).

However, this commit only introduces the state of the development branch
regarding the basic infrastructure, i.e. (mostly) `src/jace`, but leaves
out the translators that the development branch has and its tests, to
keep the PR small.