Disabled the simplify pass again in the tests.
I enabled the simplify pass in commit `411bd7bd` and it worked locally.
However, this was because I was not running it inside nox and using my own version of DaCe.
The bug in simplify was fixed in [PR#1603](https://github.com/spcl/dace/pull/1603) which was merged _after_ 16.1 was released, thus the fix is not avaliable.